+++ packages/matter.js/src/CommissioningController.ts
@@ -423,13 +423,22 @@ export class CommissioningController {
             return existingNode;
         }
 
+        const { caseAuthenticatedTags = this.#options.caseAuthenticatedTags } = connectOptions ?? {};
         const pairedNode = await PairedNode.create(
             nodeId,
             this,
             connectOptions,
             nodeIsCommissioned ? (this.#controllerInstance?.getCommissionedNodeDetails(nodeId)?.deviceData ?? {}) : {},
-            await this.createInteractionClient(nodeId, NodeDiscoveryType.None, { forcedConnection: false }), // First connect without discovery to last known address
-            async (discoveryType?: NodeDiscoveryType) => void (await controller.connect(nodeId, { discoveryType })),
+            await this.createInteractionClient(nodeId, NodeDiscoveryType.None, {
+                forcedConnection: false,
+                caseAuthenticatedTags,
+            }), // First connect without discovery to last known address
+            async (discoveryType?: NodeDiscoveryType) =>
+                void (await controller.connect(nodeId, {
+                    discoveryOptions: { discoveryType },
+                    allowUnknownPeer: false,
+                    caseAuthenticatedTags,
+                })),
             handler => this.#sessionDisconnectedHandler.set(nodeId, handler),
             controller.sessions,
             this.#crypto,
@@ -511,14 +520,22 @@ export class CommissioningController {
         discoveryType?: NodeDiscoveryType,
         options?: {
             forcedConnection?: boolean;
+            caseAuthenticatedTags?: CaseAuthenticatedTag[];
         },
     ): Promise<InteractionClient> {
         const controller = this.#assertControllerIsStarted();
-        const { forcedConnection } = options ?? {};
+        const { forcedConnection, caseAuthenticatedTags = this.#options.caseAuthenticatedTags } = options ?? {};
         if (nodeIdOrChannel instanceof MessageChannel || !forcedConnection) {
-            return controller.createInteractionClient(nodeIdOrChannel, { discoveryType });
+            return controller.createInteractionClient(nodeIdOrChannel, {
+                discoveryOptions: { discoveryType },
+                caseAuthenticatedTags,
+            });
         }
-        return controller.connect(nodeIdOrChannel, { discoveryType }, forcedConnection);
+        return controller.connect(nodeIdOrChannel, {
+            discoveryOptions: { discoveryType },
+            allowUnknownPeer: forcedConnection,
+            caseAuthenticatedTags,
+        });
     }
 
     /**